+++ .github/actions/cache-solana/action.yml
@@ -4,7 +4,7 @@ env:
 runs:
   using: composite
   steps:
-    - uses: actions/cache@v2
+    - uses: actions/cache@v3
       id: cache-solana
       with:
         path: |
