+++ packages/sdks/sunrise/src/state.ts
@@ -1,6 +1,6 @@
 import { type PublicKey } from "@solana/web3.js";
 import { type IdlAccounts, type BN } from "@coral-xyz/anchor";
-import { type SunriseBeam } from "../../types/sunrise_beam";
+import { type SunriseBeam } from "../../types/sunrise_core";
 
 /** The deserialized state for the on-chain beam account.*/
 export class StateAccount {
