+++ programs/marinade-lp-beam/Cargo.toml
@@ -19,4 +19,4 @@ default = []
 anchor-lang = '0.28.0'
 anchor-spl = '0.28.0'
 marinade-cpi = { git = "https://github.com/sunrise-stake/anchor-gen" }
-sunrise_beam = { path = "../sunrise_beam", features = ["cpi"] }
+sunrise-core = { path = "../sunrise-core", features = ["cpi"] }
