+++ programs/sunrise-core/src/instructions/burn_gsol.rs
