+++ programs/sunrise-core/src/instructions/register_beam.rs
