+++ programs/sunrise-core/src/instructions/remove_beam.rs
