+++ programs/sunrise-core/tests/register_beam.rs
