+++ programs/sunrise-core/tests/remove_beam.rs
